Move matchers to their own mod
Also do a cargo fmt pass.
